Refactor project structure and implement SysDir function for file management
