+++ go.mod
@@ -1,3 +1,3 @@
-module todo
+module sorter
 
 go 1.23.5
